+++ cishouseholds/pipeline/high_level_transformations.py
@@ -1,7 +1,9 @@
 # flake8: noqa
+import pyspark.sql.functions as F
 from pyspark.sql import DataFrame
 from pyspark.sql import functions as F
 from pyspark.sql import Window
+from pyspark.sql.dataframe import DataFrame
 
 from cishouseholds.derive import assign_age_at_date
 from cishouseholds.derive import assign_column_from_mapped_list_key
@@ -26,7 +28,9 @@ from cishouseholds.derive import assign_named_buckets
 from cishouseholds.derive import assign_outward_postcode
 from cishouseholds.derive import assign_raw_copies
 from cishouseholds.derive import assign_school_year_september_start
+from cishouseholds.derive import assign_substring
 from cishouseholds.derive import assign_taken_column
+from cishouseholds.derive import assign_test_target
 from cishouseholds.derive import assign_true_if_any
 from cishouseholds.derive import assign_unique_id_column
 from cishouseholds.derive import assign_visit_order
@@ -38,9 +42,15 @@ from cishouseholds.derive import assign_work_status_group
 from cishouseholds.derive import concat_fields_if_true
 from cishouseholds.derive import contact_known_or_suspected_covid_type
 from cishouseholds.derive import count_value_occurrences_in_column_subset_row_wise
+from cishouseholds.derive import derive_cq_pattern
+from cishouseholds.derive import derive_had_symptom_last_7days_from_digital
 from cishouseholds.derive import derive_household_been_columns
+from cishouseholds.derive import map_options_to_bool_columns
+from cishouseholds.derive import mean_across_columns
 from cishouseholds.edit import apply_value_map_multiple_columns
 from cishouseholds.edit import assign_from_map
+from cishouseholds.edit import clean_barcode
+from cishouseholds.edit import clean_barcode_simple
 from cishouseholds.edit import clean_postcode
 from cishouseholds.edit import clean_within_range
 from cishouseholds.edit import convert_null_if_not_in_list
@@ -48,9 +58,11 @@ from cishouseholds.edit import edit_to_sum_or_max_value
 from cishouseholds.edit import format_string_upper_and_clean
 from cishouseholds.edit import map_column_values_to_null
 from cishouseholds.edit import update_column_if_ref_in_list
+from cishouseholds.edit import update_column_in_time_window
 from cishouseholds.edit import update_column_values_from_map
 from cishouseholds.edit import update_face_covering_outside_of_home
 from cishouseholds.edit import update_person_count_from_ages
+from cishouseholds.edit import update_strings_to_sentence_case
 from cishouseholds.edit import update_think_have_covid_symptom_any
 from cishouseholds.edit import update_to_value_if_any_not_null
 from cishouseholds.edit import update_work_facing_now_column
@@ -64,8 +76,828 @@ from cishouseholds.impute import impute_by_ordered_fill_forward
 from cishouseholds.impute import impute_latest_date_flag
 from cishouseholds.impute import impute_outside_uk_columns
 from cishouseholds.impute import impute_visit_datetime
+from cishouseholds.pipeline.high_level_transformations import transform_survey_responses_generic
 from cishouseholds.pipeline.timestamp_map import cis_digital_datetime_map
+from cishouseholds.pyspark_utils import get_or_create_spark_session
 from cishouseholds.validate_class import SparkValidate
+
+
+def transform_blood_delta(df: DataFrame) -> DataFrame:
+    """
+    Call functions to process input for blood deltas.
+    """
+    df = assign_test_target(df, "antibody_test_target", "blood_test_source_file")
+    df = assign_substring(
+        df,
+        column_name_to_assign="antibody_test_plate_common_id",
+        column_to_substring="antibody_test_plate_id",
+        start_position=5,
+        substring_length=5,
+    )
+    df = assign_unique_id_column(
+        df=df,
+        column_name_to_assign="unique_antibody_test_id",
+        concat_columns=["blood_sample_barcode", "antibody_test_plate_common_id", "antibody_test_well_id"],
+    )
+    df = clean_barcode(
+        df=df, barcode_column="blood_sample_barcode", edited_column="blood_sample_barcode_edited_in_bloods_dataset_flag"
+    )
+    return df
+
+
+def add_historical_fields(df: DataFrame):
+    """
+    Add empty values for union with historical data. Also adds constant
+    values for continuation with historical data.
+    """
+    historical_columns = {
+        "siemens_antibody_test_result_classification": "string",
+        "siemens_antibody_test_result_value": "float",
+        "antibody_test_tdi_result_value": "float",
+        "lims_id": "string",
+        "plate_storage_method": "string",
+    }
+    for column, type in historical_columns.items():
+        if column not in df.columns:
+            df = df.withColumn(column, F.lit(None).cast(type))
+    if "antibody_assay_category" not in df.columns:
+        df = assign_column_uniform_value(df, "antibody_assay_category", "Post 2021-03-01")
+    df = df.select(sorted(df.columns))
+    return df
+
+
+def add_fields(df: DataFrame):
+    """Add fields that might be missing in example data."""
+    new_columns = {
+        "antibody_test_undiluted_result_value": "float",
+        "antibody_test_bounded_result_value": "float",
+    }
+    for column, type in new_columns.items():
+        if column not in df.columns:
+            df = df.withColumn(column, F.lit(None).cast(type))
+    df = df.select(sorted(df.columns))
+    return df
+
+
+def transform_swab_delta(df: DataFrame) -> DataFrame:
+    """
+    Transform swab delta - derive new fields that do not depend on merging with survey responses.
+    """
+    spark_session = get_or_create_spark_session()
+    df = clean_barcode(
+        df=df, barcode_column="swab_sample_barcode", edited_column="swab_sample_barcode_edited_in_swab_dataset_flag"
+    )
+    df = assign_column_to_date_string(df, "pcr_result_recorded_date_string", "pcr_result_recorded_datetime")
+    df = derive_cq_pattern(
+        df, ["orf1ab_gene_pcr_cq_value", "n_gene_pcr_cq_value", "s_gene_pcr_cq_value"], spark_session
+    )
+    df = assign_unique_id_column(
+        df, "unique_pcr_test_id", ["swab_sample_barcode", "pcr_result_recorded_datetime", "cq_pattern"]
+    )
+
+    df = mean_across_columns(
+        df, "mean_pcr_cq_value", ["orf1ab_gene_pcr_cq_value", "n_gene_pcr_cq_value", "s_gene_pcr_cq_value"]
+    )
+    df = assign_isin_list(
+        df=df,
+        column_name_to_assign="one_positive_pcr_target_only",
+        reference_column="cq_pattern",
+        values_list=["N only", "OR only", "S only"],
+        true_false_values=[1, 0],
+    )
+    return df
+
+
+def transform_swab_delta_testKit(df: DataFrame):
+    df = df.drop("testKit")
+
+    return df
+
+
+def digital_specific_cleaning(df):
+    df = assign_raw_copies(
+        df,
+        [
+            "work_in_additional_paid_employment",
+            "self_isolating",
+            "illness_lasting_over_12_months",
+            "ever_smoked_regularly",
+            "currently_smokes_or_vapes",
+            "hours_a_day_with_someone_else_at_home",
+            "face_covering_work_or_education",
+        ],
+    )
+    df = apply_value_map_multiple_columns(
+        df,
+        {
+            "work_in_additional_paid_employment": {"prefer_not_to_say": None},
+            "self_isolating": {"prefer_not_to_say": None},
+            "illness_lasting_over_12_months": {"prefer_not_to_say": None},
+            "ever_smoked_regularly": {"prefer_not_to_say": None},
+            "currently_smokes_or_vapes": {"prefer_not_to_say": None},
+            "hours_a_day_with_someone_else_at_home": {"prefer_not_to_say": None},
+            "face_covering_work_or_education": {"prefer_not_to_say": None},
+        },
+    )
+
+
+def digital_specific_transformations(df):
+    df = df.withColumn("face_covering_outside_of_home", F.lit(None).cast("string"))
+    df = concat_fields_if_true(df, "think_had_covid_which_symptoms", "think_had_covid_which_symptom_", "Yes", ";")
+    df = concat_fields_if_true(df, "which_symptoms_last_7_days", "think_have_covid_symptom_", "Yes", ";")
+    df = concat_fields_if_true(df, "long_covid_symptoms", "think_have_long_covid_symptom_", "Yes", ";")
+    df = update_column_values_from_map(
+        df,
+        "survey_completion_status",
+        {
+            "In Progress": "Partially Completed",
+            "IN PROGRESS": "Partially Completed",
+            "Submitted": "Completed",
+            "SUBMITTED": "Completed",
+        },
+    )
+    df = derive_had_symptom_last_7days_from_digital(
+        df,
+        "think_have_covid_symptom_any",
+        "think_have_covid_symptom_",
+        [
+            "fever",
+            "muscle_ache",
+            "fatigue",
+            "sore_throat",
+            "cough",
+            "shortness_of_breath",
+            "headache",
+            "nausea_vomiting",
+            "abdominal_pain",
+            "diarrhoea",
+            "loss_of_taste",
+            "loss_of_smell",
+        ],
+    )
+    return df
+
+
+def transform_survey_responses_version_0_delta(df: DataFrame) -> DataFrame:
+    """
+    Call functions to process input for iqvia version 0 survey deltas.
+    """
+    df = assign_column_uniform_value(df, "survey_response_dataset_major_version", 0)
+    df = df.withColumn("sex", F.coalesce(F.col("sex"), F.col("gender"))).drop("gender")
+
+    df = map_column_values_to_null(
+        df=df,
+        value="Participant Would Not/Could Not Answer",
+        column_list=[
+            "ethnicity",
+            "work_status_v0",
+            "work_location",
+            "survey_response_type",
+            "participant_withdrawal_reason",
+            "work_not_from_home_days_per_week",
+        ],
+    )
+
+    column_editing_map = {
+        "work_location": {
+            "Both (working from home and working outside of your home)": "Both (from home and somewhere else)",
+            "Working From Home": "Working from home",
+            "Working Outside of your Home": "Working somewhere else (not your home)",
+            "Not applicable": "Not applicable, not currently working",
+        },
+        "last_covid_contact_type": {
+            "In your own household": "Living in your own home",
+            "Outside your household": "Outside your home",
+        },
+        "last_suspected_covid_contact_type": {
+            "In your own household": "Living in your own home",
+            "Outside your household": "Outside your home",
+        },
+        "other_covid_infection_test_results": {
+            "Positive": "One or more positive test(s)",
+        },
+    }
+    df = apply_value_map_multiple_columns(df, column_editing_map)
+
+    df = clean_barcode(df=df, barcode_column="swab_sample_barcode", edited_column="swab_sample_barcode_edited_flag")
+    df = clean_barcode(df=df, barcode_column="blood_sample_barcode", edited_column="blood_sample_barcode_edited_flag")
+    return df
+
+
+def clean_survey_responses_version_1(df: DataFrame) -> DataFrame:
+    df = map_column_values_to_null(
+        df=df,
+        value="Participant Would Not/Could Not Answer",
+        column_list=[
+            "ethnicity",
+            "work_sector",
+            "work_health_care_area",
+            "work_status_v1",
+            "work_location",
+            "work_direct_contact_patients_or_clients",
+            "survey_response_type",
+            "self_isolating_reason",
+            "illness_reduces_activity_or_ability",
+            "ability_to_socially_distance_at_work_or_education",
+            "transport_to_work_or_education",
+            "face_covering_outside_of_home",
+            "other_antibody_test_location",
+            "participant_withdrawal_reason",
+            "work_not_from_home_days_per_week",
+        ],
+    )
+
+    df = df.withColumn("work_main_job_changed", F.lit(None).cast("string"))
+    fill_forward_columns = [
+        "work_main_job_title",
+        "work_main_job_role",
+        "work_sector",
+        "work_sector_other",
+        "work_health_care_area",
+        "work_nursing_or_residential_care_home",
+        "work_direct_contact_patients_or_clients",
+    ]
+    df = update_to_value_if_any_not_null(
+        df=df,
+        column_name_to_assign="work_main_job_changed",
+        value_to_assign="Yes",
+        column_list=fill_forward_columns,
+    )
+    return df
+
+
+def transform_survey_responses_version_1_delta(df: DataFrame) -> DataFrame:
+    """
+    Call functions to process input for iqvia version 1 survey deltas.
+    """
+    df = assign_column_uniform_value(df, "survey_response_dataset_major_version", 1)
+
+    df = df.withColumn("work_status_v0", F.col("work_status_v1"))
+    df = df.withColumn("work_status_v2", F.col("work_status_v1"))
+
+    been_value_map = {"No, someone else in my household has": "No I haven’t, but someone else in my household has"}
+    column_editing_map = {
+        "work_status_v0": {
+            "Employed and currently working": "Employed",  # noqa: E501
+            "Employed and currently not working": "Furloughed (temporarily not working)",  # noqa: E501
+            "Self-employed and currently not working": "Furloughed (temporarily not working)",  # noqa: E501
+            "Retired": "Not working (unemployed, retired, long-term sick etc.)",  # noqa: E501
+            "Looking for paid work and able to start": "Not working (unemployed, retired, long-term sick etc.)",  # noqa: E501
+            "Not working and not looking for work": "Not working (unemployed, retired, long-term sick etc.)",  # noqa: E501
+            "Child under 5y not attending child care": "Student",  # noqa: E501
+            "Child under 5y attending child care": "Student",  # noqa: E501
+            "5y and older in full-time education": "Student",  # noqa: E501
+            "Self-employed and currently working": "Self-employed",  # noqa: E501
+        },
+        "work_status_v2": {
+            "Child under 5y not attending child care": "Child under 4-5y not attending child care",  # noqa: E501
+            "Child under 5y attending child care": "Child under 4-5y attending child care",  # noqa: E501
+            "5y and older in full-time education": "4-5y and older at school/home-school",  # noqa: E501
+        },
+        "household_been_hospital_last_28_days": been_value_map,
+        "household_been_care_home_last_28_days": been_value_map,
+        "times_outside_shopping_or_socialising_last_7_days": {
+            "None": 0,
+            "1": 1,
+            "2": 2,
+            "3": 3,
+            "4": 4,
+            "5": 5,
+            "6": 6,
+            "7 times or more": 7,
+        },
+    }
+
+    df = assign_isin_list(
+        df=df,
+        column_name_to_assign="self_isolating",
+        reference_column="self_isolating_reason",
+        values_list=[
+            "Yes, for other reasons (e.g. going into hospital, quarantining)",
+            "Yes, you have/have had symptoms",
+            "Yes, someone you live with had symptoms",
+        ],
+        true_false_values=["Yes", "No"],
+    )
+    df = apply_value_map_multiple_columns(df, column_editing_map)
+    df = clean_barcode(df=df, barcode_column="swab_sample_barcode", edited_column="swab_sample_barcode_edited_flag")
+    df = clean_barcode(df=df, barcode_column="blood_sample_barcode", edited_column="blood_sample_barcode_edited_flag")
+    return df
+
+
+def digital_specific_transformations(df: DataFrame) -> DataFrame:
+    """
+    Call functions to process digital specific variable transformations.
+    """
+    df = assign_column_uniform_value(df, "survey_response_dataset_major_version", 3)
+    df = update_strings_to_sentence_case(df, ["survey_completion_status", "survey_not_completed_reason_code"])
+    df = df.withColumn("visit_id", F.col("participant_completion_window_id"))
+    df = df.withColumn(
+        "swab_manual_entry", F.when(F.col("swab_sample_barcode_user_entered").isNull(), "No").otherwise("Yes")
+    )
+    df = df.withColumn(
+        "blood_manual_entry", F.when(F.col("blood_sample_barcode_user_entered").isNull(), "No").otherwise("Yes")
+    )
+    df = df.withColumn(
+        "visit_datetime",
+        F.coalesce(
+            F.col("swab_taken_datetime"),
+            F.col("blood_taken_datetime"),
+            F.col("survey_completed_datetime"),
+            F.col("sample_kit_dispatched_datetime"),
+        ),
+    )  # Placeholder for 2199
+    df = update_column_in_time_window(
+        df,
+        "digital_survey_collection_mode",
+        "survey_completed_datetime",
+        "Telephone",
+        ["20-05-2022T21:30:00", "26-05-2022 00:00:00"],
+    )
+    df = transform_survey_responses_generic(df)
+
+    dont_know_columns = [
+        "work_in_additional_paid_employment",
+        "work_nursing_or_residential_care_home",
+        "work_direct_contact_patients_or_clients",
+        "self_isolating",
+        "illness_lasting_over_12_months",
+        "ever_smoked_regularly",
+        "currently_smokes_or_vapes",
+        "cis_covid_vaccine_type_1",
+        "cis_covid_vaccine_type_2",
+        "cis_covid_vaccine_type_3",
+        "cis_covid_vaccine_type_4",
+        "cis_covid_vaccine_type_5",
+        "cis_covid_vaccine_type_6",
+        "other_household_member_hospital_last_28_days",
+        "other_household_member_care_home_last_28_days",
+        "hours_a_day_with_someone_else_at_home",
+        "physical_contact_under_18_years",
+        "physical_contact_18_to_69_years",
+        "physical_contact_over_70_years",
+        "social_distance_contact_under_18_years",
+        "social_distance_contact_18_to_69_years",
+        "social_distance_contact_over_70_years",
+        "times_hour_or_longer_another_home_last_7_days",
+        "times_hour_or_longer_another_person_your_home_last_7_days",
+        "times_shopping_last_7_days",
+        "times_socialising_last_7_days",
+        "face_covering_work_or_education",
+        "face_covering_other_enclosed_places",
+        "cis_covid_vaccine_type",
+    ]
+    df = assign_raw_copies(df, dont_know_columns)
+    dont_know_mapping_dict = {"Prefer not to say": None, "Don't Know": None, "I don't know the type": None}
+    df = apply_value_map_multiple_columns(
+        df,
+        {k: dont_know_mapping_dict for k in dont_know_columns},
+    )
+
+    df = df.withColumn("self_isolating_reason_digital", F.col("self_isolating_reason"))
+    df = assign_column_value_from_multiple_column_map(
+        df,
+        "self_isolating_reason",
+        [
+            ["No", ["No", None]],
+            [
+                "Yes, you have/have had symptoms",
+                ["Yes", "I have or have had symptoms of COVID-19 or a positive test"],
+            ],
+            [
+                "Yes, someone you live with had symptoms",
+                [
+                    "Yes",
+                    "I haven't had any symptoms but I live with someone who has or has had symptoms or a positive test",
+                ],
+            ],
+            [
+                "Yes, for other reasons (e.g. going into hospital, quarantining),",  # noqa: E501
+                [
+                    "Yes",
+                    "Due to increased risk of getting COVID-19 such as having been in contact with a known case or quarantining after travel abroad",  # noqa: E501
+                ],
+            ],
+            [
+                "Yes, for other reasons (e.g. going into hospital, quarantining),",  # noqa: E501
+                ["Yes", "Due to reducing my risk of getting COVID-19 such as going into hospital or shielding"],
+            ],
+        ],
+        ["self_isolating", "self_isolating_reason"],
+    )
+
+    column_list = ["work_status_digital", "work_status_employment", "work_status_unemployment", "work_status_education"]
+    df = assign_column_value_from_multiple_column_map(
+        df,
+        "work_status_v2",
+        [
+            [
+                "Employed and currently working",
+                [
+                    "Employed",
+                    "Currently working. This includes if you are on sick or other leave for less than 4 weeks",
+                    None,
+                    None,
+                ],
+            ],
+            [
+                "Employed and currently not working",
+                [
+                    "Employed",
+                    [
+                        "Currently not working -  for example on sick or other leave such as maternity or paternity for longer than 4 weeks",  # noqa: E501
+                        "Or currently not working -  for example on sick or other leave such as maternity or paternity for longer than 4 weeks?",  # noqa: E501
+                    ],
+                    None,
+                    None,
+                ],
+            ],
+            [
+                "Self-employed and currently working",
+                [
+                    "Self-employed",
+                    "Currently working. This includes if you are on sick or other leave for less than 4 weeks",
+                    None,
+                    None,
+                ],
+            ],
+            [
+                "Self-employed and currently not working",
+                [
+                    "Self-employed",
+                    [
+                        "Currently not working -  for example on sick or other leave such as maternity or paternity for longer than 4 weeks",  # noqa: E501
+                        "Or currently not working -  for example on sick or other leave such as maternity or paternity for longer than 4 weeks?",  # noqa: E501
+                    ],
+                    None,
+                    None,
+                ],
+            ],
+            [
+                "Looking for paid work and able to start",
+                [
+                    "Not in paid work. This includes being unemployed or retired or doing voluntary work",
+                    None,
+                    "Looking for paid work and able to start",
+                    None,
+                ],
+            ],
+            [
+                "Not working and not looking for work",
+                [
+                    "Not in paid work. This includes being unemployed or retired or doing voluntary work",
+                    None,
+                    "Not looking for paid work. This includes looking after the home or family or not wanting a job or being long-term sick or disabled",  # noqa: E501
+                    None,
+                ],
+            ],
+            [
+                "Retired",
+                [
+                    "Not in paid work. This includes being unemployed or retired or doing voluntary work",
+                    None,
+                    ["Retired", "Or retired?"],
+                    None,
+                ],
+            ],
+            [
+                "Child under 4-5y not attending child care",
+                [
+                    "In education",
+                    None,
+                    None,
+                    "A child below school age and not attending a nursery or pre-school or childminder",
+                ],
+            ],
+            [
+                "Child under 4-5y attending child care",
+                [
+                    "In education",
+                    None,
+                    None,
+                    "A child below school age and attending a nursery or a pre-school or childminder",
+                ],
+            ],
+            [
+                "4-5y and older at school/home-school",
+                [
+                    "In education",
+                    None,
+                    None,
+                    ["A child aged 4 or over at school", "A child aged 4 or over at home-school"],
+                ],
+            ],
+            [
+                "Attending college or FE (including if temporarily absent)",
+                [
+                    "In education",
+                    None,
+                    None,
+                    "Attending a college or other further education provider including apprenticeships",
+                ],
+            ],
+            [
+                "Attending university (including if temporarily absent)",
+                [
+                    "In education",
+                    None,
+                    None,
+                    ["Attending university", "Or attending university?"],
+                ],
+            ],
+        ],
+        column_list,
+    )
+    df = assign_column_value_from_multiple_column_map(
+        df,
+        "work_status_v0",
+        [
+            [
+                "Employed",
+                [
+                    "Employed",
+                    "Currently working. This includes if you are on sick or other leave for less than 4 weeks",
+                    None,
+                    None,
+                ],
+            ],
+            [
+                "Not working (unemployed, retired, long-term sick etc.)",
+                [
+                    "Employed",
+                    "Currently not working. This includes if you are on sick or other leave such as maternity or paternity for longer than 4 weeks",  # noqa: E501
+                    None,
+                    None,
+                ],
+            ],
+            ["Employed", ["Self-employed", None, "Looking for paid work and able to start", None]],
+            [
+                "Self-employed",
+                [
+                    "Self-employed",
+                    None,
+                    "Not looking for paid work. This includes looking after the home or family or not wanting a job or being long-term sick or disabled",  # noqa: E501
+                    None,
+                ],
+            ],
+            ["Self-employed", [None, None, None]],
+            ["Not working (unemployed, retired, long-term sick etc.)", ["Self-employed", None, None, None]],
+            [
+                "Not working (unemployed, retired, long-term sick etc.)",
+                [
+                    "Not in paid work. This includes being unemployed or doing voluntary work",
+                    None,
+                    "Looking for paid work and able to start",
+                    None,
+                ],
+            ],
+            [
+                "Not working (unemployed, retired, long-term sick etc.)",
+                [
+                    "Not in paid work. This includes being unemployed or doing voluntary work",
+                    None,
+                    "Not looking for paid work. This includes looking after the home or family or not wanting a job or being long-term sick or disabled",  # noqa: E501
+                    None,
+                ],
+            ],
+            [
+                "Not working (unemployed, retired, long-term sick etc.)",
+                ["Not in paid work. This includes being unemployed or doing voluntary work", None, "Retired", None],
+            ],
+            [
+                "Not working (unemployed, retired, long-term sick etc.)",
+                ["Not in paid work. This includes being unemployed or doing voluntary work", None, None, None],
+            ],
+            [
+                "Student",
+                [
+                    "Education",
+                    None,
+                    None,
+                    "A child below school age and not attending a nursery or pre-school or childminder",
+                ],
+            ],
+            [
+                "Student",
+                [
+                    "Education",
+                    None,
+                    None,
+                    "A child below school age and attending a nursery or pre-school or childminder",
+                ],
+            ],
+            ["Student", ["Education", None, None, "A child aged 4 or over at school"]],
+            ["Student", ["Education", None, None, "A child aged 4 or over at home-school"]],
+            [
+                "Student",
+                [
+                    "Education",
+                    None,
+                    None,
+                    "Attending a college or other further education provider including apprenticeships",
+                ],
+            ],
+        ],
+        column_list,
+    )
+    df = clean_barcode_simple(df, "swab_sample_barcode_user_entered")
+    df = clean_barcode_simple(df, "blood_sample_barcode_user_entered")
+    df = map_options_to_bool_columns(
+        df,
+        "currently_smokes_or_vapes_description",
+        {
+            "cigarettes": "smoke_cigarettes",
+            "cigars": "smokes_cigar",
+            "pipe": "smokes_pipe",
+            "vape/E-cigarettes": "smokes_vape_e_cigarettes",
+            "Hookah/shisha pipes": "smokes_hookah_shisha_pipes",
+        },
+        ";",
+    )
+    df = df.withColumn("times_outside_shopping_or_socialising_last_7_days", F.lit(None))
+    raw_copy_list = [
+        "participant_survey_status",
+        "participant_withdrawal_type",
+        "survey_response_type",
+        "work_sector",
+        "illness_reduces_activity_or_ability",
+        "ability_to_socially_distance_at_work_or_education",
+        "last_covid_contact_type",
+        "last_suspected_covid_contact_type",
+        "physical_contact_under_18_years",
+        "physical_contact_18_to_69_years",
+        "physical_contact_over_70_years",
+        "social_distance_contact_under_18_years",
+        "social_distance_contact_18_to_69_years",
+        "social_distance_contact_over_70_years",
+        "times_hour_or_longer_another_home_last_7_days",
+        "times_hour_or_longer_another_person_your_home_last_7_days",
+        "times_shopping_last_7_days",
+        "times_socialising_last_7_days",
+        "face_covering_work_or_education",
+        "face_covering_other_enclosed_places",
+        "other_covid_infection_test_results",
+        "other_antibody_test_results",
+        "cis_covid_vaccine_type",
+        "cis_covid_vaccine_number_of_doses",
+        "cis_covid_vaccine_type_1",
+        "cis_covid_vaccine_type_2",
+        "cis_covid_vaccine_type_3",
+        "cis_covid_vaccine_type_4",
+        "cis_covid_vaccine_type_5",
+        "cis_covid_vaccine_type_6",
+    ]
+    df = assign_raw_copies(df, [column for column in raw_copy_list if column in df.columns])
+    """
+    Sets categories to map for digital specific variables to Voyager 0/1/2 equivalent
+    """
+    contact_people_value_map = {
+        "1 to 5": "1-5",
+        "6 to 10": "6-10",
+        "11 to 20": "11-20",
+        "Don't know": None,
+        "Prefer not to say": None,
+    }
+    times_value_map = {
+        "1": 1,
+        "2": 2,
+        "3": 3,
+        "4": 4,
+        "5": 5,
+        "6": 6,
+        "7 times or more": 7,
+        "Don't know": None,
+        "None": 0,
+        "Prefer not to say": None,
+    }
+    vaccine_type_map = {
+        "Pfizer / BioNTech": "Pfizer/BioNTech",
+        "Oxford / AstraZeneca": "Oxford/AstraZeneca",
+        "Janssen / Johnson&Johnson": "Janssen/Johnson&Johnson",
+        "Another vaccine please specify": "Other / specify",
+        "I don't know the type": "Don't know type",
+    }
+    column_editing_map = {
+        "participant_survey_status": {"Complete": "Completed"},
+        "participant_withdrawal_type": {
+            "Withdrawn - no future linkage": "Withdrawn_no_future_linkage",
+            "Withdrawn - no future linkage or use of samples": "Withdrawn_no_future_linkage_or_use_of_samples",
+        },
+        "survey_response_type": {"First Survey": "First Visit", "Follow-up Survey": "Follow-up Visit"},
+        "voucher_type_preference": {"Letter": "Paper", "Email": "email_address"},
+        "work_sector": {
+            "Social Care": "Social care",
+            "Transport. This includes storage and logistics": "Transport (incl. storage, logistic)",
+            "Retail sector. This includes wholesale": "Retail sector (incl. wholesale)",
+            "Hospitality - for example hotels or restaurants or cafe": "Hospitality (e.g. hotel, restaurant)",
+            "Food production and agriculture. This includes farming": "Food production, agriculture, farming",
+            "Personal Services - for example hairdressers or tattooists": "Personal services (e.g. hairdressers)",
+            "Information technology and communication": "Information technology and communication",
+            "Financial services. This includes insurance": "Financial services incl. insurance",
+            "Civil Service or Local Government": "Civil service or Local Government",
+            "Arts or entertainment or recreation": "Arts,Entertainment or Recreation",
+            "Other employment sector please specify": "Other occupation sector",
+        },
+        "work_health_care_area": {
+            "Primary care - for example in a GP or dentist": "Yes, in primary care, e.g. GP, dentist",
+            "Secondary care - for example in a hospital": "Yes, in secondary care, e.g. hospital",
+            "Another type of healthcare - for example mental health services": "Yes, in other healthcare settings, e.g. mental health",  # noqa: E501
+        },
+        "illness_reduces_activity_or_ability": {
+            "Yes a little": "Yes, a little",
+            "Yes a lot": "Yes, a lot",
+        },
+        "work_location": {
+            "From home meaning in the same grounds or building as your home": "Working from home",
+            "Somewhere else meaning not at your home)": "Working somewhere else (not your home)",
+            "Both from home and work somewhere else": "Both (from home and somewhere else)",
+        },
+        "transport_to_work_or_education": {
+            "Bus or minibus or coach": "Bus, minibus, coach",
+            "Motorbike or scooter or moped": "Motorbike, scooter or moped",
+            "Taxi or minicab": "Taxi/minicab",
+            "Underground or Metro or Light Rail or Tram": "Underground, metro, light rail, tram",
+        },
+        "ability_to_socially_distance_at_work_or_education": {
+            "Difficult to maintain 2 metres apart. But you can usually be at least 1 metre away from other people": "Difficult to maintain 2m, but can be 1m",  # noqa: E501
+            "Easy to maintain 2 metres apart. It is not a problem to stay this far away from other people": "Easy to maintain 2m",  # noqa: E501
+            "Relatively easy to maintain 2 metres apart. Most of the time you can be 2 meters away from other people": "Relatively easy to maintain 2m",  # noqa: E501
+            "Very difficult to be more than 1m away as your work means you are in close contact with others on a regular basis": "Very difficult to be more than 1m away",  # noqa: E501
+        },
+        "last_covid_contact_type": {
+            "Someone I live with": "Living in your own home",
+            "Someone I do not live with": "Outside your home",
+        },
+        "last_suspected_covid_contact_type": {
+            "Someone I live with": "Living in your own home",
+            "Someone I do not live with": "Outside your home",
+        },
+        "physical_contact_under_18_years": contact_people_value_map,
+        "physical_contact_18_to_69_years": contact_people_value_map,
+        "physical_contact_over_70_years": contact_people_value_map,
+        "social_distance_contact_under_18_years": contact_people_value_map,
+        "social_distance_contact_18_to_69_years": contact_people_value_map,
+        "social_distance_contact_over_70_years": contact_people_value_map,
+        "times_hour_or_longer_another_home_last_7_days": times_value_map,
+        "times_hour_or_longer_another_person_your_home_last_7_days": times_value_map,
+        "times_shopping_last_7_days": times_value_map,
+        "times_socialising_last_7_days": times_value_map,
+        "face_covering_work_or_education": {
+            "Prefer not to say": None,
+            "Yes sometimes": "Yes, sometimes",
+            "Yes always": "Yes, always",
+            "I am not going to my place of work or education": "Not going to place of work or education",
+            "I cover my face for other reasons - for example for religious or cultural reasons": "My face is already covered",  # noqa: E501
+        },
+        "face_covering_other_enclosed_places": {
+            "Prefer not to say": None,
+            "Yes sometimes": "Yes, sometimes",
+            "Yes always": "Yes, always",
+            "I am not going to other enclosed public spaces or using public transport": "Not going to other enclosed public spaces or using public transport",  # noqa: E501
+            "I cover my face for other reasons - for example for religious or cultural reasons": "My face is already covered",  # noqa: E501
+        },
+        "other_covid_infection_test_results": {
+            "All tests failed": "All Tests failed",
+            "One or more tests were negative and none were positive": "Any tests negative, but none positive",
+            "One or more tests were positive": "One or more positive test(s)",
+        },
+        "other_antibody_test_results": {
+            "All tests failed": "All Tests failed",
+            "One or more tests were negative for antibodies and none were positive": "Any tests negative, but none positive",  # noqa: E501
+            "One or more tests were positive for antibodies": "One or more positive test(s)",
+        },
+        "cis_covid_vaccine_type": vaccine_type_map,
+        "cis_covid_vaccine_number_of_doses": {
+            "1 dose": "1",
+            "2 doses": "2",
+            "3 doses": "3 or more",
+            "4 doses": "3 or more",
+            "5 doses": "3 or more",
+            "6 doses or more": "3 or more",
+        },
+        "cis_covid_vaccine_type_1": vaccine_type_map,
+        "cis_covid_vaccine_type_2": vaccine_type_map,
+        "cis_covid_vaccine_type_3": vaccine_type_map,
+        "cis_covid_vaccine_type_4": vaccine_type_map,
+        "cis_covid_vaccine_type_5": vaccine_type_map,
+        "cis_covid_vaccine_type_6": vaccine_type_map,
+    }
+    df = apply_value_map_multiple_columns(df, column_editing_map)
+
+    df = edit_to_sum_or_max_value(
+        df=df,
+        column_name_to_assign="times_outside_shopping_or_socialising_last_7_days",
+        columns_to_sum=[
+            "times_shopping_last_7_days",
+            "times_socialising_last_7_days",
+        ],
+        max_value=7,
+    )
+    df = df.withColumn(
+        "work_not_from_home_days_per_week",
+        F.greatest("work_not_from_home_days_per_week", "education_in_person_days_per_week"),
+    )
+    return df
 
 
 def transform_survey_responses_generic(df: DataFrame) -> DataFrame:
