+++ cishouseholds/pipeline/input_file_stages.py
@@ -2,16 +2,16 @@ from cishouseholds.pipeline.blood_delta_ETL import add_historical_fields
 from cishouseholds.pipeline.blood_delta_ETL import transform_blood_delta
 from cishouseholds.pipeline.cast_columns_from_string_map import survey_response_cast_to_double
 from cishouseholds.pipeline.cast_columns_from_string_map import survey_response_cisd_cast_to_double
+from cishouseholds.pipeline.high_level_transformations import clean_survey_responses_version_2
+from cishouseholds.pipeline.high_level_transformations import derive_additional_v1_2_columns
+from cishouseholds.pipeline.high_level_transformations import transform_survey_responses_generic
+from cishouseholds.pipeline.high_level_transformations import transform_survey_responses_version_2_delta
 from cishouseholds.pipeline.historical_blood_ETL import add_fields
 from cishouseholds.pipeline.input_variable_names import column_name_maps
 from cishouseholds.pipeline.pipeline_stages import generate_input_processing_function
 from cishouseholds.pipeline.survey_responses_version_0_ETL import transform_survey_responses_version_0_delta
 from cishouseholds.pipeline.survey_responses_version_1_ETL import clean_survey_responses_version_1
 from cishouseholds.pipeline.survey_responses_version_1_ETL import transform_survey_responses_version_1_delta
-from cishouseholds.pipeline.survey_responses_version_2_ETL import clean_survey_responses_version_2
-from cishouseholds.pipeline.survey_responses_version_2_ETL import derive_additional_v1_2_columns
-from cishouseholds.pipeline.survey_responses_version_2_ETL import transform_survey_responses_generic
-from cishouseholds.pipeline.survey_responses_version_2_ETL import transform_survey_responses_version_2_delta
 from cishouseholds.pipeline.survey_responses_version_digital_ETL import digital_specific_transformations
 from cishouseholds.pipeline.swab_delta_ETL import transform_swab_delta
 from cishouseholds.pipeline.swab_delta_ETL_testKit import transform_swab_delta_testKit
