--- cishouseholds/pipeline/swab_delta_ETL_testKit.py
@@ -1,7 +1,0 @@
-from pyspark.sql import DataFrame
-
-
-def transform_swab_delta_testKit(df: DataFrame):
-    df = df.drop("testKit")
-
-    return df
